Updated magnalister build version 8572 to latest modified-shop compatible build version 8643
git-svn-id: https://svn.modified-shop.org/svn/modified-shop/trunk@15227 79ac0487-9261-4f26-a34d-858dee0dcce0

+++ includes/external/magnalister/MagnaUpdater.php
@@ -267,7 +267,7 @@ class MagnaUpdater {
 		/* a call to set_time_limit restarts the timeout counter from zero. */
 		#$this->log(__METHOD__.' :: '.print_r(func_get_args(), true)."\n");
 		
-		@set_time_limit(ini_get('max_execution_time'));
+		@set_time_limit((int)ini_get('max_execution_time')); // (int) - type must be strict in PHP >8.1
 		
 		if ($destination === false) {
 			$destination = $this->paths['DIR_MAGNALISTER'].$name;
